Modify how the traces get added to the stats
Traces are only added if the energy consumption is non-zero. This
prevents the data from being dirty of measurements, where the RAPL
counter did not manage to generate a new value. This way measurements
done on very fast instructions (such as `add`) are not handicapped by
the slow update rate of RAPL.